safari-534.53-branch/LayoutTests/platform/gtk/fast/history/self-is-visited-expected.txt

CONSOLE MESSAGE: line 14: TypeError: Result of expression 'firstStyle' [undefined] is not an object.
These two links should be different colors (green and orange):

One Two
PASS
